Move Search from Activity to Fragment that belongs to MainActivity.
This greatly improves handling of search results because now the event
listeners of the MainActivity can be used for the search result too.

+++ detlef/src/at/ac/tuwien/detlef/db/EpisodeDAO.java
@@ -73,7 +73,7 @@ public interface EpisodeDAO {
      *            The ID of the episode to fetch.
      * @return The episode with the given ID, if it exists. Otherwise null.
      */
-    Episode getEpisode(int id);
+    Episode getEpisode(long id);
 
     /**
      * Gets an episode by the given guid or the url (both should be unique).
